Working full, probably without mistakes

+++ AISDE_nr1/AISDE_nr1/Packet.cs
@@ -10,7 +10,7 @@ namespace AISDE_nr1
     {
         public int size;
         public int priority;
-        public int adding_time;
+        public double adding_time;
 
         public Packet()
         {
